Moving implementations from plugin to api project
