+++ src/RepoM.Api/Ordering/Az/AzComparer.cs
@@ -1,6 +1,8 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Az;
+namespace RepoM.Api.Ordering.Az;
 
 using System;
+using RepoM.Core.Plugin;
+using RepoM.Core.Plugin.RepositoryOrdering;
 
 public class AzComparer : IRepositoryComparer
 {
