+++ src/RepoM.Api/Ordering/Composition/CompositionComparerConfigurationV1.cs
@@ -1,9 +1,9 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Sum;
+namespace RepoM.Api.Ordering.Composition;
 
 using System.Collections.Generic;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
 
-public class SumComparerConfigurationV1 : IRepositoriesComparerConfiguration
+public class CompositionComparerConfigurationV1 : IRepositoriesComparerConfiguration
 {
     public List<IRepositoriesComparerConfiguration> Comparers { get; set; } = new();
 }
