+++ src/RepoM.Api/Ordering/Label/TagScorerConfigurationV1.cs
@@ -1,4 +1,4 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Label;
+namespace RepoM.Api.Ordering.Label;
 
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
