+++ src/RepoM.Api/Ordering/Label/TagScorerConfigurationV1Registration.cs
@@ -1,4 +1,4 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Label;
+namespace RepoM.Api.Ordering.Label;
 
 using System;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
