+++ src/RepoM.Api/Ordering/Score/ScoreComparer.cs
@@ -1,6 +1,8 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Score;
+namespace RepoM.Api.Ordering.Score;
 
 using System;
+using RepoM.Core.Plugin;
+using RepoM.Core.Plugin.RepositoryOrdering;
 
 public class ScoreComparer : IRepositoryComparer
 {
