+++ src/RepoM.Api/Ordering/Score/ScoreComparerConfigurationV1Registration.cs
@@ -1,4 +1,4 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Score;
+namespace RepoM.Api.Ordering.Score;
 
 using System;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
