+++ src/RepoM.Api/Ordering/Sum/SumComparerConfigurationV1.cs
@@ -1,9 +1,9 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Composition;
+namespace RepoM.Api.Ordering.Sum;
 
 using System.Collections.Generic;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
 
-public class CompositionComparerConfigurationV1 : IRepositoriesComparerConfiguration
+public class SumComparerConfigurationV1 : IRepositoriesComparerConfiguration
 {
     public List<IRepositoriesComparerConfiguration> Comparers { get; set; } = new();
 }
