+++ src/RepoM.Api/Ordering/Sum/SumCompositionComparer.cs
@@ -1,7 +1,9 @@
-namespace RepoM.Core.Plugin.RepositoryOrdering.Implementations.Sum;
+namespace RepoM.Api.Ordering.Sum;
 
 using System.Collections.Generic;
 using System.Linq;
+using RepoM.Core.Plugin;
+using RepoM.Core.Plugin.RepositoryOrdering;
 
 public class SumCompositionComparer : IRepositoryComparer
 {
