+++ src/RepoM.Api/RepositoryActions/Executors/NullActionExecutor.cs
@@ -1,4 +1,4 @@
-namespace RepoM.Api.RepositoryActions.Executors.Null;
+namespace RepoM.Api.RepositoryActions.Executors;
 
 using RepoM.Core.Plugin.RepositoryActions;
 using RepoM.Core.Plugin.RepositoryActions.Actions;
