+++ tests/Yaml.Poc.Tests/DI/Bootstrapper.cs
@@ -3,14 +3,14 @@ namespace Yaml.Poc.Tests.DI;
 using System;
 using System.Threading.Tasks;
 using RepoM.Api.Git;
+using RepoM.Api.Ordering.Az;
+using RepoM.Api.Ordering.Composition;
 using RepoM.Api.Ordering.IsPinned;
+using RepoM.Api.Ordering.Label;
+using RepoM.Api.Ordering.Score;
+using RepoM.Api.Ordering.Sum;
 using RepoM.Core.Plugin.RepositoryOrdering;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Az;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Composition;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Label;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Score;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Sum;
 using SimpleInjector;
 using Container = SimpleInjector.Container;
