+++ tests/Yaml.Poc.Tests/YamlTests.cs
@@ -3,13 +3,13 @@ namespace Yaml.Poc.Tests;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using RepoM.Api.Ordering.Az;
 using RepoM.Api.Ordering.IsPinned;
+using RepoM.Api.Ordering.Score;
+using RepoM.Api.Ordering.Sum;
 using RepoM.Core.Plugin;
 using RepoM.Core.Plugin.RepositoryOrdering;
 using RepoM.Core.Plugin.RepositoryOrdering.Configuration;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Az;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Score;
-using RepoM.Core.Plugin.RepositoryOrdering.Implementations.Sum;
 using SimpleInjector;
 using VerifyXunit;
 using Xunit;
